cleanup and move to metabase folder
