+++ openshift/scripts/metabase/metabase_restore_common.sh
