+++ openshift/scripts/metabase/oc_restore_metabase_dashboards.sh
